shortener/app/requirements.txt
flask
gunicorn
redispy
